+++ CHANGELOG.md
@@ -7,6 +7,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
+- `anane view` can now return factory activity with `--activity`
+- `anane view` also accepts `-n` for all sub-option (list TFs, list regions and activity)
+
+### Changed
+- `ananse network` uses slightly less memory, and should now keep going unless all memory is used.
+
+### Fixed
+- `ananse influence` now throws a descriptive error if source and target networks are identical
+- `ananse influence` now uses scipy >=1.9, which fixes an error in `mannwhitneyu(method="auto")`
+- uploaded correct non-human CAGE-seq models
+
+## [0.4.0] - 2022-06-02
+
+### Added
+- CAGE-seq support!
+  - for more details, check out https://github.com/vanheeringen-lab/ANANSE-CAGE
 - `ananse.binding` can now optionally accept a raw counts file instead of BAM files.
   - `ananse.binding` now has a (case-insensitive) `--columns` argument to filter the counts table by.
 - `ananse.view` now optionally accepts `regions` and `tfs`, to filter the output.